two things...
one, only add a page to the page history if we change pages. Previsouly
we were adding it if we reshowed pages. So you might have to "back" twice
to go to the previous page.

bugzilla: 132672

Second, call the module obj's "noNetwork" method if it has one and there
is no network.
bugzilla: 133219