firstboot/src/firstbootWindow.py: ~440-460
The hooks for a box with no network support are getting ran before we
decided if we should just skip a module entirely or not. For the
RHN modules, the hook for no network support includes writing out the
register_remind file, which then causes the applet to remind the user
to register.

firstboot fix is to check if we skip the module entirely (as we do with
RHN stuff on fedora) before we run the no network stuff.

Also, the applet should probably require the register_remind file and
a config pointing at RHN before nagging.

bugzilla: 158095