Reorganize to provide a firstboot module.
Both firstboot and anaconda provide a constants.py file, and the sys.path
trickery we were doing in firstboot modules was causing problems.  There's
no reason we need to continue doing this when firstboot now provides a
real API and can also provide a python module.  Most files needed to be
changed to handling import differently.