Clear the user entry text fields (#736193)
The user is created immediately when pressed forward,
so if we go back, and the text fields are already filled
with the old values, after we press forward again,
firstboot tries to create the user again which results in an error.
If the text fields are cleared out, you can skip creating
the user, because you already have one created.