Let WM place the window (#821077)
Some window managers take a bit of time to be ready to use. With KDE
this was causing the screen to have its top left corner centered instead
of filling the screen depending on who won the race.

This fix doesn't try to center the window, it sets the size to match the
screen and lets the WM decide where to put it.

This has been tested by forcing the race condition with a 10s sleep
before starting the wm. This consistently reproduced the problem and the
fix consistently works.